Adding BaseQuery that provides concrete implementation for Query
